+++ packages/components/src/components/Link/Link.tsx
@@ -12,7 +12,7 @@ import styles from "./Link.module.scss";
 import clsx from "clsx";
 import type { FlowComponentProps } from "@/lib/componentFactory/flowComponent";
 import { flowComponent } from "@/lib/componentFactory/flowComponent";
-import type { PropsWithClassName } from "@/lib/types/props";
+import type { AlphaColor, PropsWithClassName } from "@/lib/types/props";
 import { linkContext } from "@/components/Link/context";
 import { LinkIcon } from "@/components/Link/components/LinkIcon";
 
@@ -26,7 +26,7 @@ export interface LinkProps
   /** An alternative link component. */
   linkComponent?: ComponentType<Omit<ComponentProps<"a">, "ref">>;
   /** The color of the link. @default "primary" */
-  color?: "primary" | "dark" | "light";
+  color?: "primary" | AlphaColor;
   "aria-current"?: string;
   slot?: string;
   /** The whiteSpace css value of the element. */